+++ changelog.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.2.7](https://github.com/syphon-org/syphon/releases/tag/0.2.6) - 2021-12-27
+
+
+### Added
+- removes exif information for image uploads
+
+### Fixed
+- fixed sending images from the quick select menu
+- fixed inviting users to chats
+- 
+
+
 ## [0.2.6](https://github.com/syphon-org/syphon/releases/tag/0.2.6) - 2021-12-27
 
 ### Fixed